refactor(desktop): rename agent config files to agent-config family
Mechanical rename only — zero behavior change. Establishes one naming
family for the canonical agent configuration surface ahead of the
flag-reduction and per-agent-adoption work:

- GlobalAgentConfigFields.tsx    -> AgentConfigFields.tsx
  (canonical fields — not global-only; definition + instance scopes
  will consume it next)
- GlobalAgentConfigEditor.tsx    -> AgentDefaultsEditor.tsx
  (the wrapper that IS about defaults keeps the defaults name)
- GlobalAgentConfigSettingsCard  -> AgentDefaultsSettingsCard
- AgentAiDefaultsDialog.tsx      -> AgentDefaultsDialog.tsx
- personaProviderModelFields.tsx -> agentConfigControls.tsx
- personaDialogPickers.tsx       -> agentConfigOptions.tsx (+ test file)

Deliberately NOT renamed:
- GlobalAgentConfig data type, tauriGlobalAgentConfig.ts, and the
  get/set_global_agent_config Tauri commands — the persistence layer
  really is global; renaming would cross the IPC/Rust boundary
- data-testids (global-agent-*) — four e2e specs depend on them,
  including an onboarding spec
- DefaultConfigStep.tsx — onboarding file, import lines updated only
