+++ desktop/src/features/agents/ui/PersonaAdvancedFields.tsx
@@ -9,7 +9,7 @@ import {
   PERSONA_FIELD_CONTROL_CLASS,
   PERSONA_FIELD_SHELL_CLASS,
   PERSONA_LABEL_OPTIONAL_CLASS,
-} from "./personaDialogPickers";
+} from "./agentConfigOptions";
 
 export function PersonaAdvancedFields({
   behaviorDraft,
