+++ desktop/src/features/agents/ui/PersonaDropdownField.tsx
@@ -13,7 +13,7 @@ import {
   type PersonaDropdownOption,
   PERSONA_FIELD_CONTROL_CLASS,
   PERSONA_FIELD_SHELL_CLASS,
-} from "./personaDialogPickers";
+} from "./agentConfigOptions";
 
 export function PersonaDropdownField({
   contentClassName,
