+++ desktop/src/features/agents/ui/PersonaModelField.tsx
@@ -10,7 +10,7 @@ import {
   PERSONA_FIELD_CONTROL_CLASS,
   PERSONA_FIELD_SHELL_CLASS,
   PERSONA_LABEL_OPTIONAL_CLASS,
-} from "./personaDialogPickers";
+} from "./agentConfigOptions";
 
 type PersonaModelFieldProps = {
   disabled: boolean;
