+++ desktop/src/features/agents/ui/PersonaProviderApiKeyField.tsx
@@ -3,11 +3,11 @@ import { Eye, EyeOff } from "lucide-react";
 
 import { cn } from "@/shared/lib/cn";
 import { Input } from "@/shared/ui/input";
-import { RequiredFieldLabel } from "./personaProviderModelFields";
+import { RequiredFieldLabel } from "./agentConfigControls";
 import {
   PERSONA_FIELD_CONTROL_CLASS,
   PERSONA_FIELD_SHELL_CLASS,
-} from "./personaDialogPickers";
+} from "./agentConfigOptions";
 
 /**
  * Top-level API key pseudo-field for providers that require a secret
