+++ desktop/src/features/agents/ui/buzzAgentConfig.test.mjs
@@ -578,7 +578,7 @@ test("provider id matching is case-insensitive", () => {
 // ---------------------------------------------------------------------------
 
 // Tests below verify the validValues membership check that the auto-clear
-// useEffect in BuzzAgentModelTuningFields and GlobalAgentConfigSettingsCard
+// useEffect in BuzzAgentModelTuningFields and AgentDefaultsSettingsCard
 // use to decide whether to reset the current effort to Inherit.
 
 test("effort max is invalid for OpenAI (should trigger auto-clear)", () => {
