+++ desktop/src/features/agents/useGlobalAgentConfig.ts
@@ -31,7 +31,7 @@ export function useGlobalAgentConfig(): {
     queryKey: globalAgentConfigQueryKey,
     queryFn: getGlobalAgentConfig,
     // Config is only mutated via setGlobalAgentConfig — treat as stable until
-    // explicitly invalidated by GlobalAgentConfigSettingsCard after a save.
+    // explicitly invalidated by AgentDefaultsSettingsCard after a save.
     staleTime: Number.POSITIVE_INFINITY,
     // Never show a stale empty flash while a background refetch runs.
     placeholderData: EMPTY_CONFIG,
